+++ src/main/java/com/futurcollars/lesson8/zad2/Rectangle.java
@@ -1,22 +1,22 @@
 package com.futurcollars.lesson8.zad2;
 
-public class Rectangle extends Shape  {
+public class Rectangle extends Shape {
 
     private int sideA;
-    private  int sideB;
+    private int sideB;
 
     public Rectangle(int sideA, int sideB) {
-        this.sideA=sideA;
-        this.sideB=sideB;
+        this.sideA = sideA;
+        this.sideB = sideB;
     }
 
     @Override
-   public int getArea(){
+    public double getArea() {
         return sideA * sideB;
     }
 
     @Override
-   public int getPerimeter(){
+    public double getPerimeter() {
         return 2 * sideA + 2 * sideB;
     }
 }
